requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.7.0
board==1.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-apscheduler==0.6.2
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-ninja==0.22.2
django-phone-field==1.8.1
django-ratelimit==4.1.0
django-seed==0.3.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Faker==19.3.1
filelock==3.12.2
frozenlist==1.4.0
future==0.18.3
geoip2==4.7.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
maxminddb==2.4.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.19
platformdirs==3.10.0
protobuf==4.24.2
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
self==2020.12.3
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
toposort==1.10
typing_extensions==4.7.1
tzlocal==5.0.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.4
user-agents==2.2.0
virtualenv==20.24.3
yarl==1.9.2
